+++ build.ps1
@@ -1,6 +1,6 @@
 param(
     [string] $VersionPrefix = "4.11.4",
-    [string] $VersionSuffix = "beta6"
+    [string] $VersionSuffix = "morph4"
 )
 
 $buildArgs = @(
