+++ src/SettingsFileUpdater/TerrariaHost/NpcData.cs
@@ -1,0 +1,14 @@
+﻿namespace SettingsFileUpdater.TerrariaHost
+{
+    public class NpcData
+    {
+        public int Id { get; set; }
+        public string FullName { get; set; }
+        public string Name { get; set; }
+        public string BestiaryId { get; set; }
+        public bool CanTalk { get; set; }
+        public bool IsCritter { get; set; }
+        public bool IsKillCredit { get; set; }
+        public int BestiaryDisplayIndex { get; set; }
+    }
+}
