+++ src/TEdit/View/CreativePowers.xaml
@@ -53,20 +53,20 @@ xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 
             <CheckBox IsChecked="{Binding Path=EnableDifficulty, Mode=TwoWay}" Content="{x:Static p:Language.tool_creative_difficulty}" VerticalAlignment="Center" Grid.Row="12" Grid.Column="1" />
             <DockPanel>
-                <TextBlock Text="{Binding Difficulty, StringFormat={}{0:0}}" Width="30" DockPanel.Dock="Right" />
-                <Slider Value="{Binding Difficulty, Mode=TwoWay}" VerticalAlignment="Center" Minimum="0" Maximum="1" SmallChange=".1" />
+                <TextBlock Text="{Binding DifficultyUI, StringFormat={}{0:0.0}}" Width="30" DockPanel.Dock="Right" />
+                <Slider Value="{Binding Difficulty, Mode=TwoWay}" VerticalAlignment="Center" Minimum="0" Maximum="1" SmallChange=".3333333" />
             </DockPanel>
 
             <CheckBox IsChecked="{Binding Path=EnableSpawnRate, Mode=TwoWay}" Content="{x:Static p:Language.tool_creative_spawn_rate}" VerticalAlignment="Center" Grid.Row="12" Grid.Column="1" />
             <DockPanel>
-                <TextBlock Text="{Binding SpawnRate, StringFormat={}{0:0}}" Width="30" DockPanel.Dock="Right" />
-                <Slider Value="{Binding SpawnRate, Mode=TwoWay}" VerticalAlignment="Center" Minimum="0" Maximum="1" SmallChange=".1" />
+                <TextBlock Text="{Binding SpawnRateUI, StringFormat={}{0:0.0}}" Width="30" DockPanel.Dock="Right" />
+                <Slider Value="{Binding SpawnRate, Mode=TwoWay}" VerticalAlignment="Center" Minimum="0" Maximum="1" SmallChange=".3333333" />
             </DockPanel>
 
             <CheckBox IsChecked="{Binding Path=EnableTimeSpeed, Mode=TwoWay}" Content="{x:Static p:Language.tool_creative_time_speed}" VerticalAlignment="Center" Grid.Row="12" Grid.Column="1" />
             <DockPanel>
-                <TextBlock Text="{Binding TimeSpeed, StringFormat={}{0:0}}" Width="30" DockPanel.Dock="Right" />
-                <Slider Value="{Binding TimeSpeed, Mode=TwoWay}" VerticalAlignment="Center" Minimum="0" Maximum="1" SmallChange=".1" />
+                <TextBlock Text="{Binding TimeSpeedUI, StringFormat={}{0:0}}" Width="30" DockPanel.Dock="Right" />
+                <Slider Value="{Binding TimeSpeed, Mode=TwoWay}" VerticalAlignment="Center" Minimum="0" Maximum="1" SmallChange=".3333333" />
             </DockPanel>
 
         </tedit:AutoGrid>
