+++ src/TEdit/View/PaintModeView.xaml
@@ -3,7 +3,8 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" xmlns:tedit="clr-namespace:TEdit.UI.Xaml"
-             xmlns:editor="clr-namespace:TEdit.Editor" xmlns:Terraria="clr-namespace:TEdit.Terraria"
+             xmlns:editor="clr-namespace:TEdit.Editor" 
+             xmlns:Terraria="clr-namespace:TEdit.Terraria"
              mc:Ignorable="d" 
              xmlns:p="clr-namespace:TEdit.Properties"
